Remove `position: absolute` and z-index from #p-logo
Follows-up (r2881, b52a2a1567). Logical backport of Vector
patch Ifb99ff36e3a (T170053).

The logo was given a higher z-index than the content, presumably
with the intent to avoid user-generated content from being able
to obscure the logo. However, setting z-index here is redundant
given #p-logo (in #column-one) is after #column (in #column-content)
which already has `position: relative` and z-index set to give it its
own stacking context that can never exceed its stack (except for
weird things - T40848).

If this z-index was needed, then it would indeed only work if
position is set to relative or absolute. Picking "absolute" in
that case would cause the logo to be taken out of the layout flow,
hence the padding on #column-one.

Remove it all, in favour of natural flow. Keep the 5px distance
by using margin-bottom (160-155=5).

Commit is a visual no-op (pixel-for-pixel identical).

Bug: T170053
Change-Id: I156c54e1dd08a1c848d1c5ec9a5a9c9f85d8680a