When the personal bar doesn't fit, make overflowing links usable
When the personal menu bar is wider than the screen (browser window),
it gets cut off on the left (and the overflowing links slide under the
logo and then off the screen, making it impossible to click them).
This is caused by the `float: right` rule added in 5a7e620. To avoid
this, add `max-width: 100%`, which results in the menu being cut off
on the right again (which lets the user view the overflowing links by
horizontal scrolling).

Bug: T226875
Change-Id: Id4186215cc069f283f8ebeca0db587e1f1369062