BaseTemplate::getFooterIcons is deprecated
For callers calling `$this->getFooterIcons( 'icononly' );`
`$this->get( 'footericons' )` can be used instead.

Note the output HTML should be identical.
BaseTemplate::getFooterIcon
will unset any items that are not arrays, but upstream changes
have made this unsupported. Using SkinTemplate::getFooterIcon
will also ensure that all icons have width and heights if not
defined which is better for performance.

Bug: T267447
Change-Id: Ia5278884a10173a509b7ae50b8407774c5d8719b